+++ client/src/test/java/replicant/ConnectorTest.java
@@ -184,7 +184,7 @@ public class ConnectorTest
 
     safeAction( connector::connect );
 
-    verify( connector.getTransport() ).connect( any( Transport.OnConnect.class ), any( Transport.OnError.class ) );
+    verify( connector.getTransport() ).connect( any( Transport.OnConnect.class ) );
 
     assertEquals( connector.getState(), ConnectorState.CONNECTING );
   }
@@ -202,7 +202,7 @@ public class ConnectorTest
     final IllegalStateException exception = new IllegalStateException();
     doAnswer( i -> {
       throw exception;
-    } ).when( connector.getTransport() ).connect( any( Transport.OnConnect.class ), any( Transport.OnError.class ) );
+    } ).when( connector.getTransport() ).connect( any( Transport.OnConnect.class ) );
 
     final IllegalStateException actual =
       expectThrows( IllegalStateException.class, () -> safeAction( connector::connect ) );