+++ client/src/test/java/replicant/ReplicantRuntimeTest.java
@@ -144,7 +144,7 @@ public class ReplicantRuntimeTest
     runtime.activate();
     schedulerLock1.dispose();
     verify( service1.getTransport(), times( 1 ) )
-      .connect( any( Transport.OnConnect.class ), any( Transport.OnError.class ) );
+      .connect( any( Transport.OnConnect.class ) );
 
     reset( service1.getTransport() );
 
@@ -156,15 +156,15 @@ public class ReplicantRuntimeTest
     entry1.getRateLimiter().fillBucket();
     runtime.activate();
     schedulerLock2.dispose();
-    verify( service1.getTransport(), never() ).connect( any( Transport.OnConnect.class ),
-                                                        any( Transport.OnError.class ) );
+    verify( service1.getTransport(), never() ).connect( any( Transport.OnConnect.class )
+    );
 
     runtime.deactivate();
     safeAction( () -> service1.setState( ConnectorState.DISCONNECTING ) );
     entry1.getRateLimiter().fillBucket();
     runtime.activate();
     verify( service1.getTransport(), never() )
-      .connect( any( Transport.OnConnect.class ), any( Transport.OnError.class ) );
+      .connect( any( Transport.OnConnect.class ) );
 
     // set service state to connected so no action required
 
@@ -176,7 +176,7 @@ public class ReplicantRuntimeTest
     runtime.activate();
     schedulerLock3.dispose();
     verify( service1.getTransport(), never() )
-      .connect( any( Transport.OnConnect.class ), any( Transport.OnError.class ) );
+      .connect( any( Transport.OnConnect.class ) );
 
     // set service state to disconnected but rate limit it
 
@@ -187,7 +187,7 @@ public class ReplicantRuntimeTest
     runtime.activate();
     schedulerLock4.dispose();
     verify( service1.getTransport(), never() )
-      .connect( any( Transport.OnConnect.class ), any( Transport.OnError.class ) );
+      .connect( any( Transport.OnConnect.class ) );
 
     // set service state to disconnected but rate limit it
 
@@ -198,7 +198,7 @@ public class ReplicantRuntimeTest
     runtime.activate();
     schedulerLock5.dispose();
     verify( service1.getTransport(), times( 1 ) )
-      .connect( any( Transport.OnConnect.class ), any( Transport.OnError.class ) );
+      .connect( any( Transport.OnConnect.class ) );
   }
 
   @Test
@@ -227,9 +227,9 @@ public class ReplicantRuntimeTest
     runtime.activate();
     schedulerLock1.dispose();
     verify( service1.getTransport(), times( 1 ) )
-      .connect( any( Transport.OnConnect.class ), any( Transport.OnError.class ) );
+      .connect( any( Transport.OnConnect.class ) );
     verify( service3.getTransport(), times( 1 ) )
-      .connect( any( Transport.OnConnect.class ), any( Transport.OnError.class ) );
+      .connect( any( Transport.OnConnect.class ) );
 
     reset( service1.getTransport() );
     reset( service3.getTransport() );
@@ -245,9 +245,9 @@ public class ReplicantRuntimeTest
     runtime.activate();
     schedulerLock2.dispose();
     verify( service1.getTransport(), never() )
-      .connect( any( Transport.OnConnect.class ), any( Transport.OnError.class ) );
+      .connect( any( Transport.OnConnect.class ) );
     verify( service3.getTransport(), never() )
-      .connect( any( Transport.OnConnect.class ), any( Transport.OnError.class ) );
+      .connect( any( Transport.OnConnect.class ) );
 
     final Disposable schedulerLock3 = pauseScheduler();
     runtime.deactivate();
@@ -258,9 +258,9 @@ public class ReplicantRuntimeTest
     runtime.activate();
     schedulerLock3.dispose();
     verify( service1.getTransport(), never() )
-      .connect( any( Transport.OnConnect.class ), any( Transport.OnError.class ) );
+      .connect( any( Transport.OnConnect.class ) );
     verify( service3.getTransport(), never() )
-      .connect( any( Transport.OnConnect.class ), any( Transport.OnError.class ) );
+      .connect( any( Transport.OnConnect.class ) );
 
     // set service state to connected so no action required
 
@@ -275,9 +275,9 @@ public class ReplicantRuntimeTest
     runtime.activate();
     schedulerLock4.dispose();
     verify( service1.getTransport(), never() )
-      .connect( any( Transport.OnConnect.class ), any( Transport.OnError.class ) );
+      .connect( any( Transport.OnConnect.class ) );
     verify( service3.getTransport(), never() )
-      .connect( any( Transport.OnConnect.class ), any( Transport.OnError.class ) );
+      .connect( any( Transport.OnConnect.class ) );
 
     // set service state to disconnected but rate limit it
 
@@ -290,9 +290,9 @@ public class ReplicantRuntimeTest
     runtime.activate();
     schedulerLock5.dispose();
     verify( service1.getTransport(), never() )
-      .connect( any( Transport.OnConnect.class ), any( Transport.OnError.class ) );
+      .connect( any( Transport.OnConnect.class ) );
     verify( service3.getTransport(), never() )
-      .connect( any( Transport.OnConnect.class ), any( Transport.OnError.class ) );
+      .connect( any( Transport.OnConnect.class ) );
 
     // set service state to disconnected but rate limit it
 
@@ -305,9 +305,9 @@ public class ReplicantRuntimeTest
     runtime.activate();
     schedulerLock6.dispose();
     verify( service1.getTransport(), times( 1 ) )
-      .connect( any( Transport.OnConnect.class ), any( Transport.OnError.class ) );
+      .connect( any( Transport.OnConnect.class ) );
     verify( service3.getTransport(), times( 1 ) )
-      .connect( any( Transport.OnConnect.class ), any( Transport.OnError.class ) );
+      .connect( any( Transport.OnConnect.class ) );
   }
 
   @Test